pyproject.toml
[project]
name = "list_master"
version = "0.1.0"
description = "Shared grocery list app — WhatsApp ingestion, LLM parsing, HTMX UI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "pydantic-settings>=2.7,<3",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "jinja2>=3.1,<4",
    "python-multipart>=0.0.18,<1",
    "httpx>=0.28,<1",
    "openai>=1.30,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.25,<1",
    "httpx>=0.28,<1",
    "ruff>=0.9,<1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
